extract post formatting tests from core
and extract helper function for normalistingCssAst
