+++ test/core-tests.js
@@ -1,21 +1,12 @@
-import css from 'css'
 import { describe, it } from 'global-mocha'
 import path from 'path'
 import penthouse from '../lib/'
 import { readFileSync as read } from 'fs'
+import normaliseCssAst from './util/normaliseCssAst'
 import chai from 'chai'
 chai.should() // binds globally on Object
 
-import ffRemover from '../lib/phantomjs/unused-fontface-remover'
-import embeddedbase64Remover from '../lib/phantomjs/embedded-base64-remover'
-import nonMatchingMediaQueryRemover from '../lib/phantomjs/non-matching-media-query-remover'
-
 process.setMaxListeners(0)
-
-// because dont want to fail tests on white space differences
-function normalisedCssAst (cssString) {
-  return css.parse(css.stringify(css.parse(cssString), { compress: true }))
-}
 
 describe('penthouse core tests', function () {
   var page1cssPath = path.join(__dirname, 'static-server', 'page1.css'),
@@ -39,8 +30,8 @@ describe('penthouse core tests', function () {
         done(err)
       }
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(yeomanExpectedCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(yeomanExpectedCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -58,8 +49,8 @@ describe('penthouse core tests', function () {
       css: pusedoRemainCssFilePath
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(pusedoRemainCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(pusedoRemainCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -97,8 +88,8 @@ describe('penthouse core tests', function () {
       css: atRuleCase0RemainCssFilePath
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(atRuleCase0RemainCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(atRuleCase0RemainCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -120,8 +111,8 @@ describe('penthouse core tests', function () {
       css: atRuleCase1RemainCssFilePath
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(atRuleCase1RemainCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(atRuleCase1RemainCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -163,8 +154,8 @@ describe('penthouse core tests', function () {
       strict: true
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(atRuleCase3RemainCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(atRuleCase3RemainCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -202,8 +193,8 @@ describe('penthouse core tests', function () {
       css: clearSelfRemainCssFilePath
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(clearSelfRemainCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(clearSelfRemainCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
@@ -225,68 +216,13 @@ describe('penthouse core tests', function () {
       ]
     }, function (err, result) {
       try {
-        var resultAst = normalisedCssAst(result)
-        var expectedAst = normalisedCssAst(forceIncludeCss)
+        var resultAst = normaliseCssAst(result)
+        var expectedAst = normaliseCssAst(forceIncludeCss)
         resultAst.should.eql(expectedAst)
         done()
       } catch (ex) {
         done(ex)
       }
     })
-  })
-
-  /* non core (non breaking) functionality tests */
-  it('should remove empty rules', function (done) {
-    var emptyRemoveCssFilePath = path.join(__dirname, 'static-server', 'empty-rules--remove.css')
-
-    penthouse({
-      url: page1,
-      css: emptyRemoveCssFilePath
-    }, function (err, result) {
-      try {
-        result.trim().should.equal('')
-        done()
-      } catch (ex) {
-        done(ex)
-      }
-    })
-  })
-
-  it('should remove embedded base64', function (done) {
-    const originalCss = read(path.join(__dirname, 'static-server', 'embedded-base64--remove.css')).toString()
-    // NOTE: penthouse's default max uri length is 1000.
-    const result = embeddedbase64Remover(originalCss, 250)
-    try {
-      result.trim().should.equal('@media (min-width: 10px) {\n\n}')
-      done()
-    } catch (ex) {
-      done(ex)
-    }
-  })
-
-  it('should remove @fontface rule, because it is not used', function (done) {
-    var fontFaceRemoveCssFilePath = path.join(__dirname, 'static-server', 'fontface--remove.css'),
-      fontFaceRemoveCss = read(fontFaceRemoveCssFilePath).toString()
-
-    var result = ffRemover(fontFaceRemoveCss)
-
-    try {
-      var resultAst = normalisedCssAst(result)
-      var expectedAst = normalisedCssAst(fontFaceRemoveCss)
-      resultAst.stylesheet.rules.should.have.length.lessThan(expectedAst.stylesheet.rules.length)
-      done()
-    } catch (ex) {
-      done(ex)
-    }
-  })
-
-  it('should remove non matching media queries', function (done) {
-    const originalCss = read(path.join(__dirname, 'static-server', 'non-matching-mq--remove.css'), 'utf8')
-    const defaultViewportRules = nonMatchingMediaQueryRemover(css.parse(originalCss).stylesheet.rules, 1300, 900)
-    defaultViewportRules.should.have.length(1)
-
-    const smallViewportRules = nonMatchingMediaQueryRemover(css.parse(originalCss).stylesheet.rules, 600, 600)
-    smallViewportRules.should.have.length(0)
-    done()
   })
 })
