+++ test/post-formatting-tests.js
@@ -1,0 +1,74 @@
+import css from 'css'
+import { describe, it } from 'global-mocha'
+import path from 'path'
+import penthouse from '../lib/'
+import { readFileSync as read } from 'fs'
+import normaliseCssAst from './util/normaliseCssAst'
+import chai from 'chai'
+chai.should() // binds globally on Object
+
+import ffRemover from '../lib/phantomjs/unused-fontface-remover'
+import embeddedbase64Remover from '../lib/phantomjs/embedded-base64-remover'
+import nonMatchingMediaQueryRemover from '../lib/phantomjs/non-matching-media-query-remover'
+
+process.setMaxListeners(0)
+
+describe('penthouse post formatting tests', function () {
+  // phantomjs takes a while to start up
+  this.timeout(5000)
+
+  it('should remove empty rules', function (done) {
+    var page1 = path.join(__dirname, 'static-server', 'page1.html')
+    var emptyRemoveCssFilePath = path.join(__dirname, 'static-server', 'empty-rules--remove.css')
+
+    penthouse({
+      url: page1,
+      css: emptyRemoveCssFilePath
+    }, function (err, result) {
+      try {
+        result.trim().should.equal('')
+        done()
+      } catch (ex) {
+        done(ex)
+      }
+    })
+  })
+
+  it('should remove embedded base64', function (done) {
+    const originalCss = read(path.join(__dirname, 'static-server', 'embedded-base64--remove.css')).toString()
+    // NOTE: penthouse's default max uri length is 1000.
+    const result = embeddedbase64Remover(originalCss, 250)
+    try {
+      result.trim().should.equal('@media (min-width: 10px) {\n\n}')
+      done()
+    } catch (ex) {
+      done(ex)
+    }
+  })
+
+  it('should remove @fontface rule, because it is not used', function (done) {
+    var fontFaceRemoveCssFilePath = path.join(__dirname, 'static-server', 'fontface--remove.css'),
+      fontFaceRemoveCss = read(fontFaceRemoveCssFilePath).toString()
+
+    var result = ffRemover(fontFaceRemoveCss)
+
+    try {
+      var resultAst = normaliseCssAst(result)
+      var expectedAst = normaliseCssAst(fontFaceRemoveCss)
+      resultAst.stylesheet.rules.should.have.length.lessThan(expectedAst.stylesheet.rules.length)
+      done()
+    } catch (ex) {
+      done(ex)
+    }
+  })
+
+  it('should remove non matching media queries', function (done) {
+    const originalCss = read(path.join(__dirname, 'static-server', 'non-matching-mq--remove.css'), 'utf8')
+    const defaultViewportRules = nonMatchingMediaQueryRemover(css.parse(originalCss).stylesheet.rules, 1300, 900)
+    defaultViewportRules.should.have.length(1)
+
+    const smallViewportRules = nonMatchingMediaQueryRemover(css.parse(originalCss).stylesheet.rules, 600, 600)
+    smallViewportRules.should.have.length(0)
+    done()
+  })
+})
